+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Change log for SharePointDsc
 
 ## Unreleased
-
+* Added Example of how to build a DSC Configuration that will take an XML file
+  from AutoSPInstaller
 * Updated documentation in regards to guidance on installing binaries from
   network locations instead of locally
 * New resources: SPFarmPropertyBag